% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_clinical_ax_scenario}
\alias{add_clinical_ax_scenario}
\title{Add clinical assessment scenario}
\usage{
add_clinical_ax_scenario(
  partial_tb,
  base_case_tb,
  data_tb,
  filter_cdn_1L_chr,
  scenario_1L_xx = NULL,
  scenarios_xx = NULL,
  uids_xx = NULL,
  unit_1L_chr = "Appointments"
)
}
\arguments{
\item{partial_tb}{Partial (a tibble)}

\item{base_case_tb}{Base case (a tibble)}

\item{data_tb}{Data (a tibble)}

\item{filter_cdn_1L_chr}{Filter condition (a character vector of length one)}

\item{scenario_1L_xx}{Scenario length one (an output object of multiple potential types), Default: NULL}

\item{scenarios_xx}{Scenarios (an output object of multiple potential types), Default: NULL}

\item{uids_xx}{Unique identifiers (an output object of multiple potential types), Default: NULL}

\item{unit_1L_chr}{Unit (a character vector of length one), Default: 'Appointments'}
}
\value{
Partial (a tibble)
}
\description{
add_clinical_ax_scenario() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add clinical assessment scenario. The function returns Partial (a tibble).
}
\keyword{internal}
