% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_cumulatives}
\alias{add_cumulatives}
\title{Add cumulatives}
\usage{
add_cumulatives(
  data_xx,
  metrics_chr,
  arrange_by_1L_chr = character(0),
  dict_by_ctg_1L_chr = FALSE,
  group_by_1L_chr = character(0),
  prefix_1L_chr = "Cumulative",
  starting_dbl = 0
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{metrics_chr}{Metrics (a character vector)}

\item{arrange_by_1L_chr}{Arrange by (a character vector of length one), Default: character(0)}

\item{dict_by_ctg_1L_chr}{Dictionary by category (a character vector of length one), Default: FALSE}

\item{group_by_1L_chr}{Group by (a character vector of length one), Default: character(0)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Cumulative'}

\item{starting_dbl}{Starting (a double vector), Default: 0}
}
\value{
Data (an output object of multiple potential types)
}
\description{
add_cumulatives() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add cumulatives. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
