% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_cyclic_cases}
\alias{add_cyclic_cases}
\title{Add cyclic cases}
\usage{
add_cyclic_cases(
  data_tb,
  arrange_by_1L_chr = character(0),
  cycle_1L_chr = c("years", "days", "weeks"),
  date_var_1L_chr = "Date",
  date_tfmn_fn = lubridate::ymd,
  end_date_dtm = NULL,
  new_zeros_chr = character(0)
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{arrange_by_1L_chr}{Arrange by (a character vector of length one), Default: character(0)}

\item{cycle_1L_chr}{Cycle (a character vector of length one), Default: c("years", "days", "weeks")}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Date'}

\item{date_tfmn_fn}{Date transformation (a function), Default: lubridate::ymd}

\item{end_date_dtm}{End date (a date vector), Default: NULL}

\item{new_zeros_chr}{New zeros (a character vector), Default: character(0)}
}
\value{
Data (a tibble)
}
\description{
add_cyclic_cases() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add cyclic cases. The function returns Data (a tibble).
}
\keyword{internal}
