% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_date_vars}
\alias{add_date_vars}
\title{Add date variables}
\usage{
add_date_vars(df, date_var_1L_chr = "Date")
}
\arguments{
\item{df}{Data.frame (a data.frame)}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Date'}
}
\value{
Data.frame (a data.frame)
}
\description{
add_date_vars() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add date variables. The function returns Data.frame (a data.frame).
}
\keyword{internal}
