% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_disengaged}
\alias{add_disengaged}
\title{Add disengaged}
\usage{
add_disengaged(
  data_xx,
  date_1L_chr,
  category_1L_chr = "Healthcare",
  date_tfmn_fn = lubridate::ymd,
  date_var_1L_chr = "Date",
  description_1L_chr = "Disengaged",
  dict_by_ctg_1L_chr = FALSE,
  uid_1L_chr = "UID",
  var_nm_1L_chr = "Disengaged"
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{date_1L_chr}{Date (a character vector of length one)}

\item{category_1L_chr}{Category (a character vector of length one), Default: 'Healthcare'}

\item{date_tfmn_fn}{Date transformation (a function), Default: lubridate::ymd}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Date'}

\item{description_1L_chr}{Description (a character vector of length one), Default: 'Disengaged'}

\item{dict_by_ctg_1L_chr}{Dictionary by category (a character vector of length one), Default: FALSE}

\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'UID'}

\item{var_nm_1L_chr}{Variable name (a character vector of length one), Default: 'Disengaged'}
}
\value{
Data (an output object of multiple potential types)
}
\description{
add_disengaged() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add disengaged. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
