% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_episodes}
\alias{add_episodes}
\title{Add episodes}
\usage{
add_episodes(
  data_xx,
  separation_after_dbl,
  index_1L_int = integer(0),
  active_var_1L_chr = "Active",
  activity_var_1L_chr = "Activity",
  date_tfmn_fn = lubridate::ymd,
  date_var_1L_chr = "Date",
  end_date_dtm = NULL,
  episode_var_1L_chr = "Episodes",
  episodes_vars_chr = character(0),
  exclude_chr = "Duration",
  fiscal_start_1L_int = 7L,
  metrics_chr = make_metric_vars(),
  prefix_1L_chr = "Cumulative",
  separations_var_1L_chr = "Separations",
  temporal_vars_chr = make_temporal_vars(),
  uid_1L_chr = "UID",
  unit_1L_chr = "month"
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{separation_after_dbl}{Separation after (a double vector)}

\item{index_1L_int}{Index (an integer vector of length one), Default: integer(0)}

\item{active_var_1L_chr}{Active variable (a character vector of length one), Default: 'Active'}

\item{activity_var_1L_chr}{Activity variable (a character vector of length one), Default: 'Activity'}

\item{date_tfmn_fn}{Date transformation (a function), Default: lubridate::ymd}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Date'}

\item{end_date_dtm}{End date (a date vector), Default: NULL}

\item{episode_var_1L_chr}{Episode variable (a character vector of length one), Default: 'Episodes'}

\item{episodes_vars_chr}{Episodes variables (a character vector), Default: character(0)}

\item{exclude_chr}{Exclude (a character vector), Default: 'Duration'}

\item{fiscal_start_1L_int}{Fiscal start (an integer vector of length one), Default: 7}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Cumulative'}

\item{separations_var_1L_chr}{Separations variable (a character vector of length one), Default: 'Separations'}

\item{temporal_vars_chr}{Temporal variables (a character vector), Default: make_temporal_vars()}

\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'UID'}

\item{unit_1L_chr}{Unit (a character vector of length one), Default: 'month'}
}
\value{
Data (an output object of multiple potential types)
}
\description{
add_episodes() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add episodes. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
