% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_fabels}
\alias{add_fabels}
\title{Add fabels}
\usage{
add_fabels(ts_models_ls, data_xx = NULL, periods_1L_int = integer(0))
}
\arguments{
\item{ts_models_ls}{Time series models (a list)}

\item{data_xx}{Data (an output object of multiple potential types), Default: NULL}

\item{periods_1L_int}{Periods (an integer vector of length one), Default: integer(0)}
}
\value{
Time series models (a list)
}
\description{
add_fabels() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add fabels. The function returns Time series models (a list).
}
\keyword{internal}
