% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_period}
\alias{add_period}
\title{Add period}
\usage{
add_period(
  data_xx,
  period_ctg_1L_chr = "Temporal",
  period_var_1L_chr = "Period",
  tenure_var_1L_chr = "Tenure"
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{period_ctg_1L_chr}{Period category (a character vector of length one), Default: 'Temporal'}

\item{period_var_1L_chr}{Period variable (a character vector of length one), Default: 'Period'}

\item{tenure_var_1L_chr}{Tenure variable (a character vector of length one), Default: 'Tenure'}
}
\value{
Data (an output object of multiple potential types)
}
\description{
add_period() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add period. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
