% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_retainers}
\alias{add_retainers}
\title{Add retainers}
\usage{
add_retainers(
  data_tb,
  datasets_ls,
  cost_var_1L_chr = "Retainer amount",
  date_var_1L_chr = "Retainer date"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{datasets_ls}{Datasets (a list)}

\item{cost_var_1L_chr}{Cost variable (a character vector of length one), Default: 'Retainer amount'}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Retainer date'}
}
\value{
No return value, called for side effects.
}
\description{
add_retainers() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add retainers. The function is called for its side effects and does not return a value.
}
\keyword{internal}
