% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sampled_imputations}
\alias{add_sampled_imputations}
\title{Add sampled imputations}
\usage{
add_sampled_imputations(
  data_xx,
  groupings_chr,
  exclude_chr = character(0),
  minimum_1L_int = 10L
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{groupings_chr}{Groupings (a character vector)}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}

\item{minimum_1L_int}{Minimum (an integer vector of length one), Default: 10}
}
\value{
No return value, called for side effects.
}
\description{
add_sampled_imputations() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add sampled imputations. The function is called for its side effects and does not return a value.
}
\keyword{internal}
