% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sampled_records}
\alias{add_sampled_records}
\title{Add sampled records}
\usage{
add_sampled_records(ds_tb, seed_lup, uid_var_nm_1L_chr = "Client ID")
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{seed_lup}{Seed (a lookup table)}

\item{uid_var_nm_1L_chr}{Unique identifier variable name (a character vector of length one), Default: 'Client ID'}
}
\value{
Dataset (a tibble)
}
\description{
add_sampled_records() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add sampled records. The function returns Dataset (a tibble).
}
\keyword{internal}
