% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sampled_variable}
\alias{add_sampled_variable}
\title{Add sampled variable}
\usage{
add_sampled_variable(
  clients_tb,
  seed_ds_tb,
  shares_dbl,
  prefix_1L_chr = "Client_",
  scale_1L_dbl = 100,
  uid_var_nm_1L_chr = "Client ID",
  var_nm_1L_chr = "Sex"
)
}
\arguments{
\item{clients_tb}{Clients (a tibble)}

\item{seed_ds_tb}{Seed dataset (a tibble)}

\item{shares_dbl}{Shares (a double vector)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Client_'}

\item{scale_1L_dbl}{Scale (a double vector of length one), Default: 100}

\item{uid_var_nm_1L_chr}{Unique identifier variable name (a character vector of length one), Default: 'Client ID'}

\item{var_nm_1L_chr}{Variable name (a character vector of length one), Default: 'Sex'}
}
\value{
Clients (a tibble)
}
\description{
add_sampled_variable() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add sampled variable. The function returns Clients (a tibble).
}
\keyword{internal}
