% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_scenario_activity_cost}
\alias{add_scenario_activity_cost}
\title{Add scenario activity cost}
\usage{
add_scenario_activity_cost(
  partial_tb,
  base_case_tb,
  unit_1L_chr = "Appointments"
)
}
\arguments{
\item{partial_tb}{Partial (a tibble)}

\item{base_case_tb}{Base case (a tibble)}

\item{unit_1L_chr}{Unit (a character vector of length one), Default: 'Appointments'}
}
\value{
Partial (a tibble)
}
\description{
add_scenario_activity_cost() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add scenario activity cost. The function returns Partial (a tibble).
}
\keyword{internal}
