% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_temporal_vars}
\alias{add_temporal_vars}
\title{Add temporal variables}
\usage{
add_temporal_vars(
  data_tb,
  date_var_1L_chr = "Date",
  fiscal_start_1L_int = 7L,
  temporal_vars_chr = make_temporal_vars()
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Date'}

\item{fiscal_start_1L_int}{Fiscal start (an integer vector of length one), Default: 7}

\item{temporal_vars_chr}{Temporal variables (a character vector), Default: make_temporal_vars()}
}
\value{
Data (a tibble)
}
\description{
add_temporal_vars() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add temporal variables. The function returns Data (a tibble).
}
\keyword{internal}
