% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_tenure}
\alias{add_tenure}
\title{Add tenure}
\usage{
add_tenure(
  data_tb,
  date_var_1L_chr = "Date",
  tenure_var_1L_chr = "Tenure",
  uid_var_1L_chr = "UID",
  unit_1L_chr = "year"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Date'}

\item{tenure_var_1L_chr}{Tenure variable (a character vector of length one), Default: 'Tenure'}

\item{uid_var_1L_chr}{Unique identifier variable (a character vector of length one), Default: 'UID'}

\item{unit_1L_chr}{Unit (a character vector of length one), Default: 'year'}
}
\value{
Data (a tibble)
}
\description{
add_tenure() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add tenure. The function returns Data (a tibble).
}
\keyword{internal}
