% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_builders.R
\name{apply_layout_to_steps}
\alias{apply_layout_to_steps}
\title{Apply board layout to steps}
\usage{
apply_layout_to_steps(steps, layout)
}
\arguments{
\item{steps}{List of \code{"serious_step"} objects.}

\item{layout}{Data frame with columns \code{id}, \code{x}, and \code{y}.}
}
\value{
Updated list of \code{"serious_step"} objects.
}
\description{
Apply board layout to steps
}
\keyword{internal}
