% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_registry.R
\name{available_capsules}
\alias{available_capsules}
\title{List available SeRiouS capsules}
\usage{
available_capsules()
}
\value{
A data frame with available capsule metadata.
}
\description{
List available SeRiouS capsules
}
