% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_autocorrelations}
\alias{calculate_autocorrelations}
\title{Calculate autocorrelations}
\usage{
calculate_autocorrelations(
  data_xx,
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly"),
  key_totals_ls = NULL,
  key_vars_chr = character(0),
  max_1L_int = NULL,
  metrics_chr = make_metric_vars()
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly")}

\item{key_totals_ls}{Key totals (a list), Default: NULL}

\item{key_vars_chr}{Key variables (a character vector), Default: character(0)}

\item{max_1L_int}{Maximum (an integer vector of length one), Default: NULL}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}
}
\value{
Autocorrelations (a list)
}
\description{
calculate_autocorrelations() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate autocorrelations. The function returns Autocorrelations (a list).
}
\keyword{internal}
