% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_running_totals}
\alias{calculate_running_totals}
\title{Calculate running totals}
\usage{
calculate_running_totals(values_num)
}
\arguments{
\item{values_num}{PARAM_DESCRIPTION}
}
\value{
Totals (a numeric)
}
\description{
calculate_running_totals() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate running totals. The function is called for its side effects and does not return a value.
}
\keyword{internal}
