% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_cells.R
\name{capsule_cells}
\alias{capsule_cells}
\title{List cells in a capsule folder}
\usage{
capsule_cells(path)
}
\arguments{
\item{path}{Path to a capsule folder.}
}
\value{
A data frame describing the cells.
}
\description{
List cells in a capsule folder
}
\seealso{
Other cell-based capsule API: 
\code{\link{capsule_add_cell}()},
\code{\link{capsule_connect_cells}()},
\code{\link{capsule_disconnect_cells}()},
\code{\link{capsule_get_cell}()},
\code{\link{capsule_move_cell}()},
\code{\link{capsule_set_cell_unlock}()},
\code{\link{capsule_update_cell}()}
}
\concept{cell-based capsule API}
