% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_cells.R
\name{capsule_connect_cells}
\alias{capsule_connect_cells}
\title{Connect two cells}
\usage{
capsule_connect_cells(path, from, to)
}
\arguments{
\item{path}{Path to a capsule folder.}

\item{from}{Source cell id.}

\item{to}{Target cell id.}
}
\value{
Invisibly returns the updated source cell.
}
\description{
Connect two cells
}
