% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_demo_iris.R
\name{capsule_demo_iris}
\alias{capsule_demo_iris}
\title{Demo iris capsule}
\usage{
capsule_demo_iris()
}
\value{
A \code{"learning_capsule"} object.
}
\description{
This capsule is a minimal example showing how to create a linear
SeRiouS learning capsule with console output, a plot and questions.
}
