% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_demo_pca.R
\name{capsule_demo_pca}
\alias{capsule_demo_pca}
\title{Demo PCA capsule}
\usage{
capsule_demo_pca()
}
\value{
A \code{"learning_capsule"} object.
}
\description{
This capsule is a pedagogical example showing how to build a
SeRiouS learning capsule around Principal Component Analysis.
}
