% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_cells.R
\name{capsule_disconnect_cells}
\alias{capsule_disconnect_cells}
\title{Disconnect two cells}
\usage{
capsule_disconnect_cells(path, from, to)
}
\arguments{
\item{path}{Path to a capsule folder.}

\item{from}{Source cell id.}

\item{to}{Target cell id.}
}
\value{
Invisibly returns the updated source cell.
}
\description{
Disconnect two cells
}
\seealso{
Other cell-based capsule API: 
\code{\link{capsule_add_cell}()},
\code{\link{capsule_cells}()},
\code{\link{capsule_connect_cells}()},
\code{\link{capsule_get_cell}()},
\code{\link{capsule_move_cell}()},
\code{\link{capsule_set_cell_unlock}()},
\code{\link{capsule_update_cell}()}
}
\concept{cell-based capsule API}
