% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_cells.R
\name{capsule_move_cell}
\alias{capsule_move_cell}
\title{Move a cell on the board}
\usage{
capsule_move_cell(path, id, x, y)
}
\arguments{
\item{path}{Path to a capsule folder.}

\item{id}{Cell id.}

\item{x, y}{Numeric coordinates.}
}
\value{
Invisibly returns the updated cell.
}
\description{
Move a cell on the board
}
