% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_registry.R
\name{capsule_registry}
\alias{capsule_registry}
\title{Internal capsule registry}
\usage{
capsule_registry()
}
\value{
A named list describing registered capsules.
}
\description{
Internal capsule registry
}
