% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_accessors.R
\name{capsule_step_ids}
\alias{capsule_step_ids}
\title{Get all step identifiers from a capsule}
\usage{
capsule_step_ids(capsule)
}
\arguments{
\item{capsule}{A \code{"learning_capsule"} object.}
}
\value{
A character vector of step identifiers.
}
\description{
Get all step identifiers from a capsule
}
