% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_cells.R
\name{capsule_update_cell}
\alias{capsule_update_cell}
\title{Update a cell in a capsule folder}
\usage{
capsule_update_cell(path, id, ...)
}
\arguments{
\item{path}{Path to a capsule folder.}

\item{id}{Cell id.}

\item{...}{Cell fields to update.}
}
\value{
Invisibly returns the updated cell.
}
\description{
Update a cell in a capsule folder
}
