% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_ds_from_web_zip}
\alias{get_ds_from_web_zip}
\title{Get dataset from web zip}
\usage{
get_ds_from_web_zip(
  file_1L_chr,
  url_1L_chr,
  mode_1L_chr = "wb",
  read_fn = read.csv,
  read_fn_args_ls = NULL
)
}
\arguments{
\item{file_1L_chr}{File (a character vector of length one)}

\item{url_1L_chr}{Url (a character vector of length one)}

\item{mode_1L_chr}{Mode (a character vector of length one), Default: 'wb'}

\item{read_fn}{Read (a function), Default: read.csv}

\item{read_fn_args_ls}{Read function arguments (a list), Default: NULL}
}
\value{
Dataset (an output object of multiple potential types)
}
\description{
get_ds_from_web_zip() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get dataset from web zip. The function returns Dataset (an output object of multiple potential types).
}
\keyword{internal}
