% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_index_type}
\alias{get_index_type}
\title{Get index type}
\usage{
get_index_type(data_tsb, index_1L_chr = character(0))
}
\arguments{
\item{data_tsb}{Data (a tsibble)}

\item{index_1L_chr}{Index (a character vector of length one), Default: character(0)}
}
\value{
Index type (a character vector of length one)
}
\description{
get_index_type() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get index type. The function returns Index type (a character vector of length one).
}
\keyword{internal}
