% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_metric_for_range}
\alias{get_metric_for_range}
\title{Get metric for range}
\usage{
get_metric_for_range(
  series_tsb,
  metric_1L_chr,
  date_end_dtm = NULL,
  date_start_dtm = NULL,
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal"),
  summary_fn = sum
)
}
\arguments{
\item{series_tsb}{Series (a tsibble)}

\item{metric_1L_chr}{Metric (a character vector of length one)}

\item{date_end_dtm}{Date end (a date vector), Default: NULL}

\item{date_start_dtm}{Date start (a date vector), Default: NULL}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal")}

\item{summary_fn}{Summary (a function), Default: sum}
}
\value{
Metric (a double vector of length one)
}
\description{
get_metric_for_range() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get metric for range. The function returns Metric (a double vector of length one).
}
\keyword{internal}
