% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_model_predrs}
\alias{get_model_predrs}
\title{Get model predictors}
\usage{
get_model_predrs(ts_models_ls = make_ts_models_ls())
}
\arguments{
\item{ts_models_ls}{Time series models (a list), Default: make_ts_models_ls()}
}
\value{
Model predictors (a list)
}
\description{
get_model_predrs() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get model predictors. The function returns Model predictors (a list).
}
\keyword{internal}
