% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_new_index}
\alias{get_new_index}
\title{Get new index}
\usage{
get_new_index(
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal",
    "sub", "fiscalyear", "fiscalquarter", "weekday")
)
}
\arguments{
\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal",
"sub", "fiscalyear", "fiscalquarter", "weekday")}
}
\value{
New index (a character vector of length one)
}
\description{
get_new_index() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get new index. The function returns New index (a character vector of length one).
}
\keyword{internal}
