% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_performance}
\alias{get_performance}
\title{Get performance}
\usage{
get_performance(
  ts_mdls_ls,
  data_xx,
  metric_1L_chr,
  rank_by_int = integer(0),
  statistics_chr = c("RMSE", "MAE", "MPE", "MAPE"),
  type_1L_chr = c("accuracy", "report")
)
}
\arguments{
\item{ts_mdls_ls}{Time series models (a list)}

\item{data_xx}{Data (an output object of multiple potential types)}

\item{metric_1L_chr}{Metric (a character vector of length one)}

\item{rank_by_int}{Rank by (an integer vector), Default: integer(0)}

\item{statistics_chr}{Statistics (a character vector), Default: c("RMSE", "MAE", "MPE", "MAPE")}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("accuracy", "report")}
}
\value{
Performance (a tibble)
}
\description{
get_performance() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get performance. The function returns Performance (a tibble).
}
\keyword{internal}
