% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_raw_erp_data}
\alias{get_raw_erp_data}
\title{Get raw Estimatedesident Population data}
\usage{
get_raw_erp_data(
  uid_1L_chr = "ERP_Q",
  age_chr = as.character(1:115),
  end_1L_chr = character(0),
  frequency_1L_chr = "quarterly",
  measure_chr = c("count", "change", "\%change"),
  region_chr = c("NSW", "VIC", "QLD", "SA", "WA", "TAS", "NT", "ACT", "OT", "AUS"),
  sex_chr = c("female", "male", "person"),
  start_1L_chr = character(0),
  version_1L_chr = NULL
)
}
\arguments{
\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'ERP_Q'}

\item{age_chr}{Age (a character vector), Default: as.character(1:115)}

\item{end_1L_chr}{End (a character vector of length one), Default: character(0)}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: 'quarterly'}

\item{measure_chr}{Measure (a character vector), Default: c("count", "change", "\%change")}

\item{region_chr}{Region (a character vector), Default: c("NSW", "VIC", "QLD", "SA", "WA", "TAS", "NT", "ACT", "OT",
"AUS")}

\item{sex_chr}{Sex (a character vector), Default: c("female", "male", "person")}

\item{start_1L_chr}{Start (a character vector of length one), Default: character(0)}

\item{version_1L_chr}{Version (a character vector of length one), Default: NULL}
}
\value{
Estimatedesident Population raw (a tibble)
}
\description{
get_raw_erp_data() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get raw estimatedesident population data. The function returns Estimatedesident Population raw (a tibble).
}
\keyword{internal}
