% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_start_end_dates}
\alias{get_start_end_dates}
\title{Get start end dates}
\usage{
get_start_end_dates(
  financial_yrs_chr,
  months_chr,
  multiple_pfx_1L_chr = "FYS",
  yr_pfx_1L_chr = "20",
  yr_two_months_chr = c("Jan", "Feb", "Mar", "Apr", "May", "Jun"),
  what_1L_chr = c("start", "end")
)
}
\arguments{
\item{financial_yrs_chr}{Financial years (a character vector)}

\item{months_chr}{Months (a character vector)}

\item{multiple_pfx_1L_chr}{Multiple prefix (a character vector of length one), Default: 'FYS'}

\item{yr_pfx_1L_chr}{Year prefix (a character vector of length one), Default: '20'}

\item{yr_two_months_chr}{Year two months (a character vector), Default: c("Jan", "Feb", "Mar", "Apr", "May", "Jun")}

\item{what_1L_chr}{What (a character vector of length one), Default: c("start", "end")}
}
\value{
Date (a date vector)
}
\description{
get_start_end_dates() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get start end dates. The function returns Date (a date vector).
}
\keyword{internal}
