% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_accessors.R
\name{get_step}
\alias{get_step}
\title{Retrieve one step from a capsule}
\usage{
get_step(capsule, id)
}
\arguments{
\item{capsule}{A \code{"learning_capsule"} object.}

\item{id}{Step identifier.}
}
\value{
A \code{"serious_step"} object.
}
\description{
Retrieve one step from a capsule
}
