% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{get_taidyverse_questionnaire}
\alias{get_taidyverse_questionnaire}
\title{Get the tAIdyverse questionnaire dataset}
\usage{
get_taidyverse_questionnaire()
}
\value{
A data frame.
}
\description{
Get the tAIdyverse questionnaire dataset
}
