% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_temporal_fn}
\alias{get_temporal_fn}
\title{Get temporal function}
\usage{
get_temporal_fn(
  period_1L_chr = make_temporal_vars(index_1L_chr = "Sub"),
  temporal_fns_ls = NULL,
  daily_fn = make_date_tfmn_fn(),
  monthly_fn = tsibble::yearmonth,
  fiscal_start_1L_int = 7L
)
}
\arguments{
\item{period_1L_chr}{Period (a character vector of length one), Default: make_temporal_vars(index_1L_chr = "Sub")}

\item{temporal_fns_ls}{Temporal functions (a list), Default: NULL}

\item{daily_fn}{Daily (a function), Default: make_date_tfmn_fn()}

\item{monthly_fn}{Monthly (a function), Default: tsibble::yearmonth}

\item{fiscal_start_1L_int}{Fiscal start (an integer vector of length one), Default: 7}
}
\value{
Temporal (a function)
}
\description{
get_temporal_fn() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get temporal function. The function returns Temporal (a function).
}
\keyword{internal}
