% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_tsibble}
\alias{get_tsibble}
\title{Get tsibble}
\usage{
get_tsibble(
  data_xx,
  fill_gaps_1L_lgl = FALSE,
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal"),
  key_totals_ls = NULL,
  key_vars_chr = character(0),
  metrics_chr = make_metric_vars(),
  prefix_1L_chr = "Cumulative",
  type_1L_chr = c("totals", "key", "wide", "main", "cumulative"),
  what_1L_chr = character(0)
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{fill_gaps_1L_lgl}{Fill gaps (a logical vector of length one), Default: FALSE}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal")}

\item{key_totals_ls}{Key totals (a list), Default: NULL}

\item{key_vars_chr}{Key variables (a character vector), Default: character(0)}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Cumulative'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("totals", "key", "wide", "main", "cumulative")}

\item{what_1L_chr}{What (a character vector of length one), Default: character(0)}
}
\value{
Data (a tsibble)
}
\description{
get_tsibble() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get tsibble. The function returns Data (a tsibble).
}
