% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_vars}
\alias{get_vars}
\title{Get variables}
\usage{
get_vars(
  data_df,
  activity_1L_chr = "Activity",
  appointments_var_1L_chr = "Appointments",
  cancellations_var_1L_chr = "Cancellations",
  clinical_team_1L_chr = "Clinical Team",
  clinician_1L_chr = "Clinician",
  clinician_discipline_1L_chr = "Service",
  components_chr = c("Year", "Quarter", "Week"),
  cost_var_1L_chr = "Cost",
  days_1L_chr = "Weekday",
  duration_1L_chr = "Duration",
  exclude_chr = character(0),
  group_1L_chr = character(0),
  index_1L_chr = "Date",
  referrals_var_1L_chr = "Referrals",
  referrers_1L_chr = "Referrer Role",
  severity_1L_chr = "Severity",
  team_disciplines_1L_chr = "Disciplines",
  uid_var_1L_chr = "UID",
  what_1L_chr = c("all", "clinical", "metrics", "sports", "temporal", "other")
)
}
\arguments{
\item{data_df}{Data (a data.frame)}

\item{activity_1L_chr}{Activity (a character vector of length one), Default: 'Activity'}

\item{appointments_var_1L_chr}{Appointments variable (a character vector of length one), Default: 'Appointments'}

\item{cancellations_var_1L_chr}{Cancellations variable (a character vector of length one), Default: 'Cancellations'}

\item{clinical_team_1L_chr}{Clinical team (a character vector of length one), Default: 'Clinical Team'}

\item{clinician_1L_chr}{Clinician (a character vector of length one), Default: 'Clinician'}

\item{clinician_discipline_1L_chr}{Clinician discipline (a character vector of length one), Default: 'Service'}

\item{components_chr}{Components (a character vector), Default: c("Year", "Quarter", "Week")}

\item{cost_var_1L_chr}{Cost variable (a character vector of length one), Default: 'Cost'}

\item{days_1L_chr}{Days (a character vector of length one), Default: 'Weekday'}

\item{duration_1L_chr}{Duration (a character vector of length one), Default: 'Duration'}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}

\item{group_1L_chr}{Group (a character vector of length one), Default: character(0)}

\item{index_1L_chr}{Index (a character vector of length one), Default: 'Date'}

\item{referrals_var_1L_chr}{Referrals variable (a character vector of length one), Default: 'Referrals'}

\item{referrers_1L_chr}{Referrers (a character vector of length one), Default: 'Referrer Role'}

\item{severity_1L_chr}{Severity (a character vector of length one), Default: 'Severity'}

\item{team_disciplines_1L_chr}{Team disciplines (a character vector of length one), Default: 'Disciplines'}

\item{uid_var_1L_chr}{Unique identifier variable (a character vector of length one), Default: 'UID'}

\item{what_1L_chr}{What (a character vector of length one), Default: c("all", "clinical", "metrics", "sports", "temporal", "other")}
}
\value{
Variables (a character vector)
}
\description{
get_vars() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get variables. The function returns Variables (a character vector).
}
\keyword{internal}
