% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_age_bands_lup}
\alias{make_age_bands_lup}
\title{Make age bands lookup table}
\usage{
make_age_bands_lup(
  bands_chr,
  values_ls,
  fractions_ls = NULL,
  type_1L_chr = "by_year"
)
}
\arguments{
\item{bands_chr}{Bands (a character vector)}

\item{values_ls}{Values (a list)}

\item{fractions_ls}{Fractions (a list), Default: NULL}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'by_year'}
}
\value{
Age bands (a lookup table)
}
\description{
make_age_bands_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make age bands lookup table. The function returns Age bands (a lookup table).
}
\keyword{internal}
