% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_base_case_tb}
\alias{make_base_case_tb}
\title{Make base case tibble}
\usage{
make_base_case_tb(
  forecast_tb,
  add_ls = NULL,
  scenario_1L_chr = "Base case",
  use_1L_chr = "Mean"
)
}
\arguments{
\item{forecast_tb}{Forecast (a tibble)}

\item{add_ls}{Add (a list), Default: NULL}

\item{scenario_1L_chr}{Scenario (a character vector of length one), Default: 'Base case'}

\item{use_1L_chr}{Use (a character vector of length one), Default: 'Mean'}
}
\value{
Base case (a tibble)
}
\description{
make_base_case_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make base case tibble. The function returns Base case (a tibble).
}
\keyword{internal}
