% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_correspondences}
\alias{make_correspondences}
\title{Make correspondences}
\usage{
make_correspondences(
  data_tb = NULL,
  key_1L_chr = character(0),
  min_1L_int = 3L,
  original_xx = character(0)
)
}
\arguments{
\item{data_tb}{Data (a tibble), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: character(0)}

\item{min_1L_int}{Minimum (an integer vector of length one), Default: 3}

\item{original_xx}{Original (an output object of multiple potential types), Default: character(0)}
}
\value{
X (Name correspondences lookup table)
}
\description{
make_correspondences() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make correspondences. The function is called for its side effects and does not return a value.
}
\keyword{internal}
