% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cost_tb}
\alias{make_cost_tb}
\title{Make cost tibble}
\usage{
make_cost_tb(
  data_tsb,
  cost_1L_chr = "Cost",
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal"),
  group_by_chr = character(0),
  group_fn = sum,
  metrics_chr = make_metric_vars(),
  temporal_vars_chr = make_temporal_vars(),
  unit_1L_chr = character(0),
  unit_cost_1L_chr = "UnitCost"
)
}
\arguments{
\item{data_tsb}{Data (a tsibble)}

\item{cost_1L_chr}{Cost (a character vector of length one), Default: 'Cost'}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal")}

\item{group_by_chr}{Group by (a character vector), Default: character(0)}

\item{group_fn}{Group (a function), Default: sum}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{temporal_vars_chr}{Temporal variables (a character vector), Default: make_temporal_vars()}

\item{unit_1L_chr}{Unit (a character vector of length one), Default: character(0)}

\item{unit_cost_1L_chr}{Unit cost (a character vector of length one), Default: 'UnitCost'}
}
\value{
Cost (a tibble)
}
\description{
make_cost_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cost tibble. The function returns Cost (a tibble).
}
\keyword{internal}
