% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cumulatives}
\alias{make_cumulatives}
\title{Make cumulatives}
\usage{
make_cumulatives(
  prefix_1L_chr = "Cumulative",
  separation_after_dbl = numeric(0)
)
}
\arguments{
\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Cumulative'}

\item{separation_after_dbl}{Separation after (a double vector), Default: numeric(0)}
}
\value{
Cumulatives (a character vector)
}
\description{
make_cumulatives() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cumulatives. The function returns Cumulatives (a character vector).
}
\keyword{internal}
