% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_date_tfmn_fn}
\alias{make_date_tfmn_fn}
\title{Make date transformation function}
\usage{
make_date_tfmn_fn(format_1L_chr = "\%d-\%b-\%y")
}
\arguments{
\item{format_1L_chr}{Format (a character vector of length one), Default: '\%d-\%b-\%y'}
}
\value{
Date transformation (a function)
}
\description{
make_date_tfmn_fn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make date transformation function. The function returns Date transformation (a function).
}
\keyword{internal}
