% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_duplicate_totals_tb}
\alias{make_duplicate_totals_tb}
\title{Make duplicate totals tibble}
\usage{
make_duplicate_totals_tb(
  data_tb,
  type_1L_chr = c("Category", "Subcategory", "Subdivision", "Discipline")
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("Category", "Subcategory", "Subdivision", "Discipline")}
}
\value{
Duplicate totals (a tibble)
}
\description{
make_duplicate_totals_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make duplicate totals tibble. The function returns Duplicate totals (a tibble).
}
\keyword{internal}
