% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_duplicates_ls}
\alias{make_duplicates_ls}
\title{Make duplicates list}
\usage{
make_duplicates_ls(data_tb)
}
\arguments{
\item{data_tb}{Data (a tibble)}
}
\value{
Duplicates (a list)
}
\description{
make_duplicates_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make duplicates list. The function returns Duplicates (a list).
}
\keyword{internal}
