% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_duplicates_tb}
\alias{make_duplicates_tb}
\title{Make duplicates tibble}
\usage{
make_duplicates_tb(data_tb, category_1L_chr, subcategory_1L_chr)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{category_1L_chr}{Category (a character vector of length one)}

\item{subcategory_1L_chr}{Subcategory (a character vector of length one)}
}
\value{
Duplicates (a tibble)
}
\description{
make_duplicates_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make duplicates tibble. The function returns Duplicates (a tibble).
}
\keyword{internal}
