% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_builders.R
\name{make_edges}
\alias{make_edges}
\title{Create capsule edges}
\usage{
make_edges(from, to)
}
\arguments{
\item{from}{Character vector of source step ids.}

\item{to}{Character vector of target step ids.}
}
\value{
A data frame with columns \code{from} and \code{to}.
}
\description{
Create capsule edges
}
\seealso{
Other advanced capsule builders: 
\code{\link{build_capsule}()},
\code{\link{build_linear_capsule}()},
\code{\link{make_layout}()},
\code{\link{make_sections}()},
\code{\link{make_step}()},
\code{\link{validate_capsule}()}
}
\concept{advanced capsule builders}
