% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_episodes_vars}
\alias{make_episodes_vars}
\title{Make episodes variables}
\usage{
make_episodes_vars(
  active_var_1L_chr = "Active",
  episode_var_1L_chr = "Episodes",
  flatten_1L_lgl = TRUE,
  prefix_1L_chr = "",
  separation_after_dbl = numeric(0),
  separations_var_1L_chr = "Separations",
  suffix_1L_chr = ""
)
}
\arguments{
\item{active_var_1L_chr}{Active variable (a character vector of length one), Default: 'Active'}

\item{episode_var_1L_chr}{Episode variable (a character vector of length one), Default: 'Episodes'}

\item{flatten_1L_lgl}{Flatten (a logical vector of length one), Default: TRUE}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: ''}

\item{separation_after_dbl}{Separation after (a double vector), Default: numeric(0)}

\item{separations_var_1L_chr}{Separations variable (a character vector of length one), Default: 'Separations'}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ''}
}
\value{
Episodes variables (an output object of multiple potential types)
}
\description{
make_episodes_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make episodes variables. The function returns Episodes variables (an output object of multiple potential types).
}
\keyword{internal}
