% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_erp_ds}
\alias{make_erp_ds}
\title{Make Estimatedesident Population dataset}
\usage{
make_erp_ds(
  erp_raw_tb,
  age_range_int,
  age_bands_lup = NULL,
  age_tfmn_fn = as.integer,
  as_dyad_1L_lgl = TRUE,
  frequency_1L_chr = "quarterly",
  measure_1L_chr = "count",
  select_chr = c("time_period", "age", "sex", "obs_value"),
  sex_chr = c("female", "male", "person"),
  summarise_1L_lgl = FALSE,
  var_ctg_chr = c("Temporal", "Key", "Key", "Metric")
)
}
\arguments{
\item{erp_raw_tb}{Estimatedesident Population raw (a tibble)}

\item{age_range_int}{Age range (an integer vector)}

\item{age_bands_lup}{Age bands (a lookup table), Default: NULL}

\item{age_tfmn_fn}{Age transformation (a function), Default: as.integer}

\item{as_dyad_1L_lgl}{As dyad (a logical vector of length one), Default: TRUE}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: 'quarterly'}

\item{measure_1L_chr}{Measure (a character vector of length one), Default: 'count'}

\item{select_chr}{Select (a character vector), Default: c("time_period", "age", "sex", "obs_value")}

\item{sex_chr}{Sex (a character vector), Default: c("female", "male", "person")}

\item{summarise_1L_lgl}{Summarise (a logical vector of length one), Default: FALSE}

\item{var_ctg_chr}{Variable category (a character vector), Default: c("Temporal", "Key", "Key", "Metric")}
}
\value{
Estimatedesident Population (an output object of multiple potential types)
}
\description{
make_erp_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make estimatedesident population dataset. The function returns Estimatedesident Population (an output object of multiple potential types).
}
\keyword{internal}
