% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_forecast_cost_tb}
\alias{make_forecast_cost_tb}
\title{Make forecast cost tibble}
\usage{
make_forecast_cost_tb(
  fabels_ls,
  unit_cost_1L_dbl,
  fixed_cost_1L_dbl = 0,
  what_1L_chr = "Appointments"
)
}
\arguments{
\item{fabels_ls}{Fabels (a list)}

\item{unit_cost_1L_dbl}{Unit cost (a double vector of length one)}

\item{fixed_cost_1L_dbl}{Fixed cost (a double vector of length one), Default: 0}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Appointments'}
}
\value{
Forecast cost (a tibble)
}
\description{
make_forecast_cost_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make forecast cost tibble. The function returns Forecast cost (a tibble).
}
\keyword{internal}
