% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_forecast_growth}
\alias{make_forecast_growth}
\title{Make forecast growth}
\usage{
make_forecast_growth(
  forecasts_tb,
  add_to_reference_1L_chr = character(0),
  group_by_1L_chr = character(0),
  pooled_mdl_1L_chr = "Pooled",
  reference_1L_chr = character(0),
  reference_1L_dbl = numeric(0),
  tfmn_fn = scales::percent
)
}
\arguments{
\item{forecasts_tb}{Forecasts (a tibble)}

\item{add_to_reference_1L_chr}{Add to reference (a character vector of length one), Default: character(0)}

\item{group_by_1L_chr}{Group by (a character vector of length one), Default: character(0)}

\item{pooled_mdl_1L_chr}{Pooled model (a character vector of length one), Default: 'Pooled'}

\item{reference_1L_chr}{Reference (a character vector of length one), Default: character(0)}

\item{reference_1L_dbl}{Reference (a double vector of length one), Default: numeric(0)}

\item{tfmn_fn}{Transformation (a function), Default: scales::percent}
}
\value{
Forecasts (a tibble)
}
\description{
make_forecast_growth() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make forecast growth. The function returns Forecasts (a tibble).
}
\keyword{internal}
