% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_forecasts_tb}
\alias{make_forecasts_tb}
\title{Make forecasts tibble}
\usage{
make_forecasts_tb(
  fabels_ls,
  group_by_1L_chr = character(0),
  group_fn = sum,
  summary_fn = mean,
  tfmn_args_ls = NULL,
  tfmn_fn = NULL,
  tfmn_pattern_1L_chr = "Transformed_{.col}",
  type_1L_chr = c("default", "grouped", "summary", "both"),
  what_1L_chr = "Appointments"
)
}
\arguments{
\item{fabels_ls}{Fabels (a list)}

\item{group_by_1L_chr}{Group by (a character vector of length one), Default: character(0)}

\item{group_fn}{Group (a function), Default: sum}

\item{summary_fn}{Summary (a function), Default: mean}

\item{tfmn_args_ls}{Transformation arguments (a list), Default: NULL}

\item{tfmn_fn}{Transformation (a function), Default: NULL}

\item{tfmn_pattern_1L_chr}{Transformation pattern (a character vector of length one), Default: 'Transformed_{.col}'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("default", "grouped", "summary", "both")}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Appointments'}
}
\value{
Forecast (a tibble)
}
\description{
make_forecasts_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make forecasts tibble. The function returns Forecast (a tibble).
}
\keyword{internal}
