% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_metric_vars}
\alias{make_metric_vars}
\title{Make metric variables}
\usage{
make_metric_vars(
  type_1L_chr = c("main", "all", "eoc"),
  appointments_var_1L_chr = "Appointments",
  cancellations_var_1L_chr = "Cancellations",
  cost_var_1L_chr = "Cost",
  referrals_var_1L_chr = "Referrals",
  separation_after_dbl = numeric(0)
)
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one), Default: c("main", "all", "eoc")}

\item{appointments_var_1L_chr}{Appointments variable (a character vector of length one), Default: 'Appointments'}

\item{cancellations_var_1L_chr}{Cancellations variable (a character vector of length one), Default: 'Cancellations'}

\item{cost_var_1L_chr}{Cost variable (a character vector of length one), Default: 'Cost'}

\item{referrals_var_1L_chr}{Referrals variable (a character vector of length one), Default: 'Referrals'}

\item{separation_after_dbl}{Separation after (a double vector), Default: numeric(0)}
}
\value{
Metric variables (a character vector)
}
\description{
make_metric_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make metric variables. The function returns Metric variables (a character vector).
}
\keyword{internal}
