% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_metrics_summary}
\alias{make_metrics_summary}
\title{Make metrics summary}
\usage{
make_metrics_summary(data_tb, index_1L_chr, key_vars_chr, metrics_chr)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{index_1L_chr}{Index (a character vector of length one)}

\item{key_vars_chr}{Key variables (a character vector)}

\item{metrics_chr}{Metrics (a character vector)}
}
\value{
Summary (a tibble)
}
\description{
make_metrics_summary() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make metrics summary. The function returns Summary (a tibble).
}
\keyword{internal}
