% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_new_cases_ls}
\alias{make_new_cases_ls}
\title{Make new cases list}
\usage{
make_new_cases_ls(cases_ls)
}
\arguments{
\item{cases_ls}{Cases (a list)}
}
\value{
New cases (a list)
}
\description{
make_new_cases_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make new cases list. The function returns New cases (a list).
}
\keyword{internal}
