% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_rebate_cap_params}
\alias{make_rebate_cap_params}
\title{Make rebate cap parameters}
\usage{
make_rebate_cap_params(
  data_tb,
  scenarios_xx,
  appointments_1L_chr = "Appointments",
  appointments_cdn_1L_chr = "Activity == \\"Appointment\\"",
  cancellations_cdn_1L_chr = "Activity == \\"Cancellation\\"",
  uids_xx = NULL
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{scenarios_xx}{Scenarios (an output object of multiple potential types)}

\item{appointments_1L_chr}{Appointments (a character vector of length one), Default: 'Appointments'}

\item{appointments_cdn_1L_chr}{Appointments condition (a character vector of length one), Default: 'Activity == "Appointment"'}

\item{cancellations_cdn_1L_chr}{Cancellations condition (a character vector of length one), Default: 'Activity == "Cancellation"'}

\item{uids_xx}{Unique identifiers (an output object of multiple potential types), Default: NULL}
}
\value{
Parameters (a list)
}
\description{
make_rebate_cap_params() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make rebate cap parameters. The function returns Parameters (a list).
}
\keyword{internal}
