% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_retainers}
\alias{make_retainers}
\title{Make retainers}
\usage{
make_retainers(
  retainers_tb,
  as_tsibble_1L_lgl = FALSE,
  censor_1L_lgl = TRUE,
  cumulatives_1L_lgl = FALSE,
  cost_var_1L_chr = "Retainer amount",
  data_tb = NULL,
  date_var_1L_chr = "Retainer date",
  default_1L_dbl = numeric(0),
  dyad_1L_lgl = FALSE,
  end_date_dtm = NULL,
  fill_gaps_1L_lgl = FALSE,
  fiscal_start_1L_int = 7L,
  offset_1L_int = -1,
  provider_id_1L_chr = "ProviderID",
  reset_new_1L_lgl = TRUE,
  unit_1L_chr = "days"
)
}
\arguments{
\item{retainers_tb}{Retainers (a tibble)}

\item{as_tsibble_1L_lgl}{As tsibble (a logical vector of length one), Default: FALSE}

\item{censor_1L_lgl}{Censor (a logical vector of length one), Default: TRUE}

\item{cumulatives_1L_lgl}{Cumulatives (a logical vector of length one), Default: FALSE}

\item{cost_var_1L_chr}{Cost variable (a character vector of length one), Default: 'Retainer amount'}

\item{data_tb}{Data (a tibble), Default: NULL}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Retainer date'}

\item{default_1L_dbl}{Default (a double vector of length one), Default: numeric(0)}

\item{dyad_1L_lgl}{Dyad (a logical vector of length one), Default: FALSE}

\item{end_date_dtm}{End date (a date vector), Default: NULL}

\item{fill_gaps_1L_lgl}{Fill gaps (a logical vector of length one), Default: FALSE}

\item{fiscal_start_1L_int}{Fiscal start (an integer vector of length one), Default: 7}

\item{offset_1L_int}{Offset (an integer vector of length one), Default: -1}

\item{provider_id_1L_chr}{Provider identity (a character vector of length one), Default: 'ProviderID'}

\item{reset_new_1L_lgl}{Reset new (a logical vector of length one), Default: TRUE}

\item{unit_1L_chr}{Unit (a character vector of length one), Default: 'days'}
}
\value{
Data (an output object of multiple potential types)
}
\description{
make_retainers() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make retainers. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
