% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_roll_back_lup}
\alias{make_roll_back_lup}
\title{Make roll back lookup table}
\usage{
make_roll_back_lup(data_xx, groupings_chr, values_chr, minimum_1L_int = 10L)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{groupings_chr}{Groupings (a character vector)}

\item{values_chr}{Values (a character vector)}

\item{minimum_1L_int}{Minimum (an integer vector of length one), Default: 10}
}
\value{
Roll back (a lookup table)
}
\description{
make_roll_back_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make roll back lookup table. The function returns Roll back (a lookup table).
}
\keyword{internal}
