% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_roll_backs_ls}
\alias{make_roll_backs_ls}
\title{Make roll backs list}
\usage{
make_roll_backs_ls(
  data_xx,
  groupings_chr,
  exclude_chr = character(0),
  minimum_1L_int = 10L
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{groupings_chr}{Groupings (a character vector)}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}

\item{minimum_1L_int}{Minimum (an integer vector of length one), Default: 10}
}
\value{
Roll backs (a list)
}
\description{
make_roll_backs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make roll backs list. The function returns Roll backs (a list).
}
\keyword{internal}
