% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sampled_values}
\alias{make_sampled_values}
\title{Make sampled values}
\usage{
make_sampled_values(
  roll_back_xx,
  draws_int,
  fail_with_xx = NULL,
  filter_cdn_ls = NULL,
  variable_1L_chr = character(0)
)
}
\arguments{
\item{roll_back_xx}{Roll back (an output object of multiple potential types)}

\item{draws_int}{Draws (an integer vector)}

\item{fail_with_xx}{Fail with (an output object of multiple potential types), Default: NULL}

\item{filter_cdn_ls}{Filter condition (a list), Default: NULL}

\item{variable_1L_chr}{Variable (a character vector of length one), Default: character(0)}
}
\value{
Sampled (an output object of multiple potential types)
}
\description{
make_sampled_values() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sampled values. The function returns Sampled (an output object of multiple potential types).
}
\keyword{internal}
