% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sampling_lup}
\alias{make_sampling_lup}
\title{Make sampling lookup table}
\usage{
make_sampling_lup(
  shares_dbl,
  values_xx,
  var_nm_1L_chr,
  prefix_1L_chr = "Client_",
  scale_1L_dbl = 100,
  uid_var_nm_1L_chr = "Client ID"
)
}
\arguments{
\item{shares_dbl}{Shares (a double vector)}

\item{values_xx}{Values (an output object of multiple potential types)}

\item{var_nm_1L_chr}{Variable name (a character vector of length one)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Client_'}

\item{scale_1L_dbl}{Scale (a double vector of length one), Default: 100}

\item{uid_var_nm_1L_chr}{Unique identifier variable name (a character vector of length one), Default: 'Client ID'}
}
\value{
Sampling (a lookup table)
}
\description{
make_sampling_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sampling lookup table. The function returns Sampling (a lookup table).
}
\keyword{internal}
