% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_scaled_forecasts}
\alias{make_scaled_forecasts}
\title{Make scaled forecasts}
\usage{
make_scaled_forecasts(
  forecasts_tb,
  predictors_tb,
  after_1L_chr = character(0),
  base_1L_dbl = numeric(0),
  before_1L_chr = character(0),
  bind_1L_lgl = TRUE,
  positive_1L_lgl = FALSE,
  prefix_1L_chr = "scenario_",
  reference_1L_chr = character(0),
  reference_1L_dbl = numeric(0),
  scale_fn = function(x) x * 1,
  scaled_fn_ls = NULL,
  tfmn_1_fn = identity,
  tfmn_2_fn = identity
)
}
\arguments{
\item{forecasts_tb}{Forecasts (a tibble)}

\item{predictors_tb}{Predictors (a tibble)}

\item{after_1L_chr}{After (a character vector of length one), Default: character(0)}

\item{base_1L_dbl}{Base (a double vector of length one), Default: numeric(0)}

\item{before_1L_chr}{Before (a character vector of length one), Default: character(0)}

\item{bind_1L_lgl}{Bind (a logical vector of length one), Default: TRUE}

\item{positive_1L_lgl}{Positive (a logical vector of length one), Default: FALSE}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'scenario_'}

\item{reference_1L_chr}{Reference (a character vector of length one), Default: character(0)}

\item{reference_1L_dbl}{Reference (a double vector of length one), Default: numeric(0)}

\item{scale_fn}{Scale (a function), Default: function(x) x * 1}

\item{scaled_fn_ls}{Scaled (a list of functions), Default: NULL}

\item{tfmn_1_fn}{Transformation 1 (a function), Default: identity}

\item{tfmn_2_fn}{Transformation 2 (a function), Default: identity}
}
\value{
Forecasts (a tibble)
}
\description{
make_scaled_forecasts() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make scaled forecasts. The function returns Forecasts (a tibble).
}
\keyword{internal}
