% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_scenario_forecasts}
\alias{make_scenario_forecasts}
\title{Make scenario forecasts}
\usage{
make_scenario_forecasts(
  scenario_forecasts_ls,
  what_1L_chr,
  bind_to_tb = NULL,
  bound_models_chr = character(0),
  date_end_dtm = NULL,
  date_start_dtm = NULL,
  date_var_1L_chr = "Day",
  group_by_1L_chr = character(0),
  group_fn = sum,
  positive_1L_lgl = FALSE,
  summarise_1L_lgl = FALSE,
  summary_fn = mean,
  summary_2_fn = sum,
  tfmn_args_ls = NULL,
  tfmn_fn = NULL,
  tfmn_pattern_1L_chr = "Transformed_{.col}",
  type_1L_chr = c("default", "grouped", "summary", "both"),
  predictors_chr = character(0)
)
}
\arguments{
\item{scenario_forecasts_ls}{Scenario forecasts (a list)}

\item{what_1L_chr}{What (a character vector of length one)}

\item{bind_to_tb}{Bind to (a tibble), Default: NULL}

\item{bound_models_chr}{Bound models (a character vector), Default: character(0)}

\item{date_end_dtm}{Date end (a date vector), Default: NULL}

\item{date_start_dtm}{Date start (a date vector), Default: NULL}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Day'}

\item{group_by_1L_chr}{Group by (a character vector of length one), Default: character(0)}

\item{group_fn}{Group (a function), Default: sum}

\item{positive_1L_lgl}{Positive (a logical vector of length one), Default: FALSE}

\item{summarise_1L_lgl}{Summarise (a logical vector of length one), Default: FALSE}

\item{summary_fn}{Summary (a function), Default: mean}

\item{summary_2_fn}{Summary 2 (a function), Default: sum}

\item{tfmn_args_ls}{Transformation arguments (a list), Default: NULL}

\item{tfmn_fn}{Transformation (a function), Default: NULL}

\item{tfmn_pattern_1L_chr}{Transformation pattern (a character vector of length one), Default: 'Transformed_{.col}'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("default", "grouped", "summary", "both")}

\item{predictors_chr}{Predictors (a character vector), Default: character(0)}
}
\value{
Forecasts (a tibble)
}
\description{
make_scenario_forecasts() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make scenario forecasts. The function returns Forecasts (a tibble).
}
\keyword{internal}
