% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_scenarios_tb}
\alias{make_scenarios_tb}
\title{Make scenarios tibble}
\usage{
make_scenarios_tb(
  base_case_tb,
  data_tb,
  scenarios_xx,
  add_cumulatives_1L_lgl = F,
  add_fn = NULL,
  data_filter_1L_chr = character(0),
  extra_costs_1L_chr = character(0),
  group_by_1L_chr = "UID",
  new_filter_1L_chr = character(0),
  prefix_1L_chr = "",
  suffix_1L_chr = "",
  uids_xx = NULL,
  unit_1L_chr = "Appointments"
)
}
\arguments{
\item{base_case_tb}{Base case (a tibble)}

\item{data_tb}{Data (a tibble)}

\item{scenarios_xx}{Scenarios (an output object of multiple potential types)}

\item{add_cumulatives_1L_lgl}{Add cumulatives (a logical vector of length one), Default: F}

\item{add_fn}{Add (a function), Default: NULL}

\item{data_filter_1L_chr}{Data filter (a character vector of length one), Default: character(0)}

\item{extra_costs_1L_chr}{Extra costs (a character vector of length one), Default: character(0)}

\item{group_by_1L_chr}{Group by (a character vector of length one), Default: 'UID'}

\item{new_filter_1L_chr}{New filter (a character vector of length one), Default: character(0)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: ''}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: ''}

\item{uids_xx}{Unique identifiers (an output object of multiple potential types), Default: NULL}

\item{unit_1L_chr}{Unit (a character vector of length one), Default: 'Appointments'}
}
\value{
Scenarios (a tibble)
}
\description{
make_scenarios_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make scenarios tibble. The function returns Scenarios (a tibble).
}
\keyword{internal}
