% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_service_summary}
\alias{make_service_summary}
\title{Make service summary}
\usage{
make_service_summary(
  data_xx,
  active_base_1L_chr = "Active",
  max_periods_1L_int = integer(0),
  metrics_chr = character(0),
  metric_var_1L_chr = "Appointments",
  missing_val_1L_dbl = 0,
  patterns_ls = NULL,
  period_ctg_1L_chr = "Temporal",
  period_var_1L_chr = "Period",
  prefix_1L_chr = "",
  service_var_1L_chr = "Service",
  summary_fn = sum,
  tenure_var_1L_chr = "Tenure",
  uid_1L_chr = "UID",
  update_desc_1L_lgl = TRUE,
  var_ctg_chr = "Summary"
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{active_base_1L_chr}{Active base (a character vector of length one), Default: 'Active'}

\item{max_periods_1L_int}{Maximum periods (an integer vector of length one), Default: integer(0)}

\item{metrics_chr}{Metrics (a character vector), Default: character(0)}

\item{metric_var_1L_chr}{Metric variable (a character vector of length one), Default: 'Appointments'}

\item{missing_val_1L_dbl}{Missing value (a double vector of length one), Default: 0}

\item{patterns_ls}{Patterns (a list), Default: NULL}

\item{period_ctg_1L_chr}{Period category (a character vector of length one), Default: 'Temporal'}

\item{period_var_1L_chr}{Period variable (a character vector of length one), Default: 'Period'}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: ''}

\item{service_var_1L_chr}{Service variable (a character vector of length one), Default: 'Service'}

\item{summary_fn}{Summary (a function), Default: sum}

\item{tenure_var_1L_chr}{Tenure variable (a character vector of length one), Default: 'Tenure'}

\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'UID'}

\item{update_desc_1L_lgl}{Update description (a logical vector of length one), Default: TRUE}

\item{var_ctg_chr}{Variable category (a character vector), Default: 'Summary'}
}
\value{
Data (an output object of multiple potential types)
}
\description{
make_service_summary() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make service summary. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
