% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_builders.R
\name{make_step_sequence}
\alias{make_step_sequence}
\title{Build a linear sequence of steps}
\usage{
make_step_sequence(steps, ncol = 5, snake = TRUE)
}
\arguments{
\item{steps}{List of \code{"serious_step"} objects.}

\item{ncol}{Number of columns for automatic board layout.}

\item{snake}{Logical. If \code{TRUE}, creates a snake-like board layout.}
}
\value{
A list of updated \code{"serious_step"} objects.
}
\description{
Build a linear sequence of steps
}
