% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_summary_ds}
\alias{make_summary_ds}
\title{Make summary dataset}
\usage{
make_summary_ds(
  data_xx,
  join_before_dtm = NULL,
  max_tenure_1L_dbl = numeric(0),
  prefix_1L_chr = "Cumulative",
  separation_after_dbl = numeric(0)
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{join_before_dtm}{Join before (a date vector), Default: NULL}

\item{max_tenure_1L_dbl}{Maximum tenure (a double vector of length one), Default: numeric(0)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Cumulative'}

\item{separation_after_dbl}{Separation after (a double vector), Default: numeric(0)}
}
\value{
Data (an output object of multiple potential types)
}
\description{
make_summary_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make summary dataset. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
