% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_temporal_fns}
\alias{make_temporal_fns}
\title{Make temporal functions}
\usage{
make_temporal_fns(
  daily_fn = make_date_tfmn_fn(),
  fiscal_start_1L_int = 7L,
  monthly_fn = tsibble::yearmonth,
  rename_1L_lgl = FALSE
)
}
\arguments{
\item{daily_fn}{Daily (a function), Default: make_date_tfmn_fn()}

\item{fiscal_start_1L_int}{Fiscal start (an integer vector of length one), Default: 7}

\item{monthly_fn}{Monthly (a function), Default: tsibble::yearmonth}

\item{rename_1L_lgl}{Rename (a logical vector of length one), Default: FALSE}
}
\value{
Temporal functions (a list)
}
\description{
make_temporal_fns() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make temporal functions. The function returns Temporal functions (a list).
}
\keyword{internal}
