% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_temporal_vars}
\alias{make_temporal_vars}
\title{Make temporal variables}
\usage{
make_temporal_vars(
  what_1L_chr = c("all", "core", "fiscal", "days", "cf", "cd", "fd"),
  index_1L_chr = character(0),
  components_chr = c("Day", "Week", "Month", "Quarter", "Year"),
  fiscal_chr = c("FiscalYQ", "FiscalQuarter", "FiscalYear"),
  days_1L_chr = "Weekday"
)
}
\arguments{
\item{what_1L_chr}{What (a character vector of length one), Default: c("all", "core", "fiscal", "days", "cf", "cd", "fd")}

\item{index_1L_chr}{Index (a character vector of length one), Default: character(0)}

\item{components_chr}{Components (a character vector), Default: c("Day", "Week", "Month", "Quarter", "Year")}

\item{fiscal_chr}{Fiscal (a character vector), Default: c("FiscalYQ", "FiscalQuarter", "FiscalYear")}

\item{days_1L_chr}{Days (a character vector of length one), Default: 'Weekday'}
}
\value{
Temporal variables (a character vector)
}
\description{
make_temporal_vars() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make temporal variables. The function returns Temporal variables (a character vector).
}
\keyword{internal}
