% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tfmn_args_ls}
\alias{make_tfmn_args_ls}
\title{Make transformation arguments list}
\usage{
make_tfmn_args_ls(
  fill_gaps_1L_lgl = logical(0),
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly"),
  join_to_xx = NULL,
  key_totals_ls = NULL,
  key_vars_chr = character(0),
  metrics_chr = character(0),
  type_1L_chr = c("totals", "key", "wide", "main", "cumulative"),
  what_1L_chr = character(0)
)
}
\arguments{
\item{fill_gaps_1L_lgl}{Fill gaps (a logical vector of length one), Default: logical(0)}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly")}

\item{join_to_xx}{Join to (an output object of multiple potential types), Default: NULL}

\item{key_totals_ls}{Key totals (a list), Default: NULL}

\item{key_vars_chr}{Key variables (a character vector), Default: character(0)}

\item{metrics_chr}{Metrics (a character vector), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("totals", "key", "wide", "main", "cumulative")}

\item{what_1L_chr}{What (a character vector of length one), Default: character(0)}
}
\value{
Transformation arguments (a list)
}
\description{
make_tfmn_args_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make transformation arguments list. The function returns Transformation arguments (a list).
}
\keyword{internal}
