% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_training_ds}
\alias{make_training_ds}
\title{Make training dataset}
\usage{
make_training_ds(data_tsb, index_1L_chr = "Date", test_1L_int = integer(0))
}
\arguments{
\item{data_tsb}{Data (a tsibble)}

\item{index_1L_chr}{Index (a character vector of length one), Default: 'Date'}

\item{test_1L_int}{Test (an integer vector of length one), Default: integer(0)}
}
\value{
Training (a tsibble)
}
\description{
make_training_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make training dataset. The function returns Training (a tsibble).
}
\keyword{internal}
