% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ts_models_ls}
\alias{make_ts_models_ls}
\title{Make time series models list}
\usage{
make_ts_models_ls(
  mabels_ls = list(),
  args_ls = make_tfmn_args_ls(),
  cumulatives_args_ls = make_tfmn_args_ls(),
  fabels_ls = list(),
  join_to_args_ls = make_tfmn_args_ls(),
  predictor_args_ls = make_tfmn_args_ls(),
  models_chr = character(0),
  test_1L_int = integer(0)
)
}
\arguments{
\item{mabels_ls}{Mabels (a list), Default: list()}

\item{args_ls}{Arguments (a list), Default: make_tfmn_args_ls()}

\item{cumulatives_args_ls}{Cumulatives arguments (a list), Default: make_tfmn_args_ls()}

\item{fabels_ls}{Fabels (a list), Default: list()}

\item{join_to_args_ls}{Join to arguments (a list), Default: make_tfmn_args_ls()}

\item{predictor_args_ls}{Predictor arguments (a list), Default: make_tfmn_args_ls()}

\item{models_chr}{Models (a character vector), Default: character(0)}

\item{test_1L_int}{Test (an integer vector of length one), Default: integer(0)}
}
\value{
Time series models (a list)
}
\description{
make_ts_models_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make time series models list. The function returns Time series models (a list).
}
\keyword{internal}
