% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_helpers.R
\name{make_tutorial_env}
\alias{make_tutorial_env}
\title{Create a tutorial environment for a capsule}
\usage{
make_tutorial_env(capsule, parent = globalenv())
}
\arguments{
\item{capsule}{A \code{"learning_capsule"} object.}

\item{parent}{Parent environment.}
}
\value{
An environment containing the initial capsule data.
}
\description{
Create a tutorial environment for a capsule
}
