% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plateau.R
\name{mod_plateau_ui}
\alias{mod_plateau_ui}
\title{Plateau module UI}
\usage{
mod_plateau_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
Shiny UI.
}
\description{
Plateau module UI
}
\keyword{internal}
