% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step_viewer.R
\name{mod_step_viewer_server}
\alias{mod_step_viewer_server}
\title{Step viewer module server}
\usage{
mod_step_viewer_server(id, capsule, state)
}
\arguments{
\item{id}{Module id.}

\item{capsule}{A \code{"learning_capsule"} object.}

\item{state}{Shared reactive state.}
}
\value{
A Shiny module server.
}
\description{
Step viewer module server
}
\keyword{internal}
