% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step_viewer.R
\name{mod_step_viewer_ui}
\alias{mod_step_viewer_ui}
\title{Step viewer module UI}
\usage{
mod_step_viewer_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
Shiny UI.
}
\description{
Step viewer module UI
}
