% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_autocorrelations}
\alias{plot_autocorrelations}
\title{Plot autocorrelations}
\usage{
plot_autocorrelations(
  data_xx,
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly"),
  max_1L_int = NULL,
  metrics_chr = make_metric_vars()
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly")}

\item{max_1L_int}{Maximum (an integer vector of length one), Default: NULL}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}
}
\value{
No return value, called for side effects.
}
\description{
plot_autocorrelations() is a Plot function that plots data. Specifically, this function implements an algorithm to plot autocorrelations. The function is called for its side effects and does not return a value.
}
\keyword{internal}
