% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_decomposition}
\alias{plot_decomposition}
\title{Plot decomposition}
\usage{
plot_decomposition(
  data_xx,
  colours_chr = c("gray", "#D55E00"),
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly"),
  key_totals_ls = NULL,
  key_vars_chr = character(0),
  metrics_chr = make_metric_vars(),
  what_1L_chr = c("all", "adjusted", "trend"),
  x_label_1L_chr = ""
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{colours_chr}{Colours (a character vector), Default: c("gray", "#D55E00")}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly")}

\item{key_totals_ls}{Key totals (a list), Default: NULL}

\item{key_vars_chr}{Key variables (a character vector), Default: character(0)}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{what_1L_chr}{What (a character vector of length one), Default: c("all", "adjusted", "trend")}

\item{x_label_1L_chr}{X label (a character vector of length one), Default: ''}
}
\value{
No return value, called for side effects.
}
\description{
plot_decomposition() is a Plot function that plots data. Specifically, this function implements an algorithm to plot decomposition. The function is called for its side effects and does not return a value.
}
\keyword{internal}
