% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_lags}
\alias{plot_lags}
\title{Plot lags}
\usage{
plot_lags(
  data_xx,
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly"),
  key_totals_ls = NULL,
  key_vars_chr = character(0),
  metrics_chr = make_metric_vars(),
  prefix_1L_chr = "Lagged "
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly")}

\item{key_totals_ls}{Key totals (a list), Default: NULL}

\item{key_vars_chr}{Key variables (a character vector), Default: character(0)}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Lagged '}
}
\value{
No return value, called for side effects.
}
\description{
plot_lags() is a Plot function that plots data. Specifically, this function implements an algorithm to plot lags. The function is called for its side effects and does not return a value.
}
\keyword{internal}
