% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_multiple}
\alias{plot_multiple}
\title{Plot multiple}
\usage{
plot_multiple(
  tsibbles_xx,
  by_value_1L_lgl = F,
  caption_1L_chr = "",
  clinical_vars_chr = make_clinical_vars(),
  keep_cdn_1L_chr = c("All", "Personal", "Provider", "Severity", "Sports"),
  label_x_1L_chr = "",
  label_y_1L_chr = character(0),
  metrics_chr = make_metric_vars(),
  metrics_int = 1:4,
  min_1L_int = 3L,
  prefix_1L_chr = "",
  severity_1L_int = 8L,
  shorten_1L_chr = character(0),
  shorten_1L_lgl = FALSE,
  sports_vars_chr = get_sports_vars(),
  suffix_1L_chr = character(0),
  type_1L_chr = c("main", "seasonal"),
  y_suffix_1L_chr = character(0)
)
}
\arguments{
\item{tsibbles_xx}{Tsibbles (an output object of multiple potential types)}

\item{by_value_1L_lgl}{By value (a logical vector of length one), Default: F}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: ''}

\item{clinical_vars_chr}{Clinical variables (a character vector), Default: make_clinical_vars()}

\item{keep_cdn_1L_chr}{Keep condition (a character vector of length one), Default: c("All", "Personal", "Provider", "Severity", "Sports")}

\item{label_x_1L_chr}{Label x (a character vector of length one), Default: ''}

\item{label_y_1L_chr}{Label y (a character vector of length one), Default: character(0)}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{metrics_int}{Metrics (an integer vector), Default: 1:4}

\item{min_1L_int}{Minimum (an integer vector of length one), Default: 3}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: ''}

\item{severity_1L_int}{Severity (an integer vector of length one), Default: 8}

\item{shorten_1L_chr}{Shorten (a character vector of length one), Default: character(0)}

\item{shorten_1L_lgl}{Shorten (a logical vector of length one), Default: FALSE}

\item{sports_vars_chr}{Sports variables (a character vector), Default: get_sports_vars()}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("main", "seasonal")}

\item{y_suffix_1L_chr}{Y suffix (a character vector of length one), Default: character(0)}
}
\value{
No return value, called for side effects.
}
\description{
plot_multiple() is a Plot function that plots data. Specifically, this function implements an algorithm to plot multiple. The function is called for its side effects and does not return a value.
}
\keyword{internal}
