% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_periods}
\alias{plot_periods}
\title{Plot periods}
\usage{
plot_periods(
  data_tb,
  colour_var_1L_chr,
  y_1L_chr,
  anonymous_1L_lgl = FALSE,
  colour_end_1L_chr = "#0e668b",
  end_1L_chr = "End",
  line_1L_int = 4,
  start_1L_chr = "Start",
  x_1L_chr = "Date"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{colour_var_1L_chr}{Colour variable (a character vector of length one)}

\item{y_1L_chr}{Y (a character vector of length one)}

\item{anonymous_1L_lgl}{Anonymous (a logical vector of length one), Default: FALSE}

\item{colour_end_1L_chr}{Colour end (a character vector of length one), Default: '#0e668b'}

\item{end_1L_chr}{End (a character vector of length one), Default: 'End'}

\item{line_1L_int}{Line (an integer vector of length one), Default: 4}

\item{start_1L_chr}{Start (a character vector of length one), Default: 'Start'}

\item{x_1L_chr}{X (a character vector of length one), Default: 'Date'}
}
\value{
Periods (a plot)
}
\description{
plot_periods() is a Plot function that plots data. Specifically, this function implements an algorithm to plot periods. The function returns Periods (a plot).
}
\keyword{internal}
