% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot scatter}
\usage{
plot_scatter(
  data_xx,
  axis_labels_1L_chr = c("none", "show", "inner"),
  caption_1L_chr = "",
  clinical_vars_chr = make_clinical_vars(),
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly"),
  grid_1L_lgl = F,
  keep_cdn_1L_chr = c("All", "Personal", "Provider", "Severity", "Sports"),
  key_totals_ls = NULL,
  key_vars_chr = character(0),
  metrics_chr = make_metric_vars(),
  min_1L_int = 3L,
  severity_1L_int = 8L,
  shorten_1L_chr = character(0),
  shorten_1L_lgl = F,
  sports_vars_chr = get_sports_vars(),
  type_1L_chr = c("totals", "key", "wide"),
  what_1L_chr = character(0)
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{axis_labels_1L_chr}{Axis labels (a character vector of length one), Default: c("none", "show", "inner")}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: ''}

\item{clinical_vars_chr}{Clinical variables (a character vector), Default: make_clinical_vars()}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly")}

\item{grid_1L_lgl}{Grid (a logical vector of length one), Default: F}

\item{keep_cdn_1L_chr}{Keep condition (a character vector of length one), Default: c("All", "Personal", "Provider", "Severity", "Sports")}

\item{key_totals_ls}{Key totals (a list), Default: NULL}

\item{key_vars_chr}{Key variables (a character vector), Default: character(0)}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{min_1L_int}{Minimum (an integer vector of length one), Default: 3}

\item{severity_1L_int}{Severity (an integer vector of length one), Default: 8}

\item{shorten_1L_chr}{Shorten (a character vector of length one), Default: character(0)}

\item{shorten_1L_lgl}{Shorten (a logical vector of length one), Default: F}

\item{sports_vars_chr}{Sports variables (a character vector), Default: get_sports_vars()}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("totals", "key", "wide")}

\item{what_1L_chr}{What (a character vector of length one), Default: character(0)}
}
\value{
No return value, called for side effects.
}
\description{
plot_scatter() is a Plot function that plots data. Specifically, this function implements an algorithm to plot scatter. The function is called for its side effects and does not return a value.
}
\keyword{internal}
