% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_sngl_series}
\alias{plot_sngl_series}
\title{Plot single series}
\usage{
plot_sngl_series(
  data_tsb,
  caption_1L_chr = "",
  grid_1L_lgl = FALSE,
  label_x_1L_chr = "",
  label_y_1L_chr = character(0),
  metrics_chr = make_metric_vars(),
  min_1L_int = 3L,
  prefix_1L_chr = "",
  shorten_1L_chr = character(0),
  shorten_1L_lgl = FALSE,
  suffix_1L_chr = character(0),
  type_1L_chr = c("main", "seasonal"),
  what_1L_chr = make_metric_vars(),
  y_suffix_1L_chr = character(0)
)
}
\arguments{
\item{data_tsb}{Data (a tsibble)}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: ''}

\item{grid_1L_lgl}{Grid (a logical vector of length one), Default: FALSE}

\item{label_x_1L_chr}{Label x (a character vector of length one), Default: ''}

\item{label_y_1L_chr}{Label y (a character vector of length one), Default: character(0)}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{min_1L_int}{Minimum (an integer vector of length one), Default: 3}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: ''}

\item{shorten_1L_chr}{Shorten (a character vector of length one), Default: character(0)}

\item{shorten_1L_lgl}{Shorten (a logical vector of length one), Default: FALSE}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("main", "seasonal")}

\item{what_1L_chr}{What (a character vector of length one), Default: make_metric_vars()}

\item{y_suffix_1L_chr}{Y suffix (a character vector of length one), Default: character(0)}
}
\value{
No return value, called for side effects.
}
\description{
plot_sngl_series() is a Plot function that plots data. Specifically, this function implements an algorithm to plot single series. The function is called for its side effects and does not return a value.
}
\keyword{internal}
