% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_tsibble}
\alias{plot_tsibble}
\title{Plot tsibble}
\usage{
plot_tsibble(
  series_tsb,
  auto_1L_lgl = TRUE,
  date_tfmn_fn = NULL,
  fiscal_start_1L_int = 7L,
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal",
    "sub"),
  key_totals_ls = NULL,
  key_vars_chr = character(0),
  metrics_chr = make_metric_vars(),
  prefix_1L_chr = "Cumulative",
  transform_1L_lgl = TRUE,
  type_1L_chr = c("main", "cumulative"),
  what_1L_chr = character(0)
)
}
\arguments{
\item{series_tsb}{Series (a tsibble)}

\item{auto_1L_lgl}{Automatic (a logical vector of length one), Default: TRUE}

\item{date_tfmn_fn}{Date transformation (a function), Default: NULL}

\item{fiscal_start_1L_int}{Fiscal start (an integer vector of length one), Default: 7}

\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal",
"sub")}

\item{key_totals_ls}{Key totals (a list), Default: NULL}

\item{key_vars_chr}{Key variables (a character vector), Default: character(0)}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Cumulative'}

\item{transform_1L_lgl}{Transform (a logical vector of length one), Default: TRUE}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("main", "cumulative")}

\item{what_1L_chr}{What (a character vector of length one), Default: character(0)}
}
\value{
Plot (a plot)
}
\description{
plot_tsibble() is a Plot function that plots data. Specifically, this function implements an algorithm to plot tsibble. The function returns Plot (a plot).
}
\keyword{internal}
