% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_weekdays}
\alias{plot_weekdays}
\title{Plot weekdays}
\usage{
plot_weekdays(dss_ls, metrics_chr = make_metric_vars(), what_1L_chr = "Winter")
}
\arguments{
\item{dss_ls}{Datasets (a list)}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Winter'}
}
\value{
No return value, called for side effects.
}
\description{
plot_weekdays() is a Plot function that plots data. Specifically, this function implements an algorithm to plot weekdays. The function is called for its side effects and does not return a value.
}
\keyword{internal}
