% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_cells.R
\name{print.serious_cell}
\alias{print.serious_cell}
\title{Print a SeRiouS cell}
\usage{
\method{print}{serious_cell}(x, ...)
}
\arguments{
\item{x}{A \code{"serious_cell"} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a SeRiouS cell
}
