% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_dir.R
\name{run_capsule_dir}
\alias{run_capsule_dir}
\title{Run a SeRiouS capsule from a folder}
\usage{
run_capsule_dir(path)
}
\arguments{
\item{path}{Path to a capsule folder.}
}
\value{
Launches a Shiny application.
}
\description{
Run a SeRiouS capsule from a folder
}
