% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{run_learning_capsule}
\alias{run_learning_capsule}
\title{Run a SeRiouS learning capsule}
\usage{
run_learning_capsule(capsule, ...)
}
\arguments{
\item{capsule}{A \code{"learning_capsule"} object or the id of a registered capsule.}

\item{...}{Additional arguments passed to the Shiny app.}
}
\value{
Launches a Shiny application.
}
\description{
Run a SeRiouS learning capsule
}
