% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_registry.R
\name{serious_internal_capsule_constructor}
\alias{serious_internal_capsule_constructor}
\title{Create a constructor for an internal cell-based capsule}
\usage{
serious_internal_capsule_constructor(id)
}
\arguments{
\item{id}{Internal capsule id.}
}
\value{
A function returning a \code{"learning_capsule"} object.
}
\description{
Create a constructor for an internal cell-based capsule
}
