% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_registry.R
\name{serious_internal_capsule_dir}
\alias{serious_internal_capsule_dir}
\title{Find an internal capsule directory}
\usage{
serious_internal_capsule_dir(id)
}
\arguments{
\item{id}{Internal capsule id.}
}
\value{
A path, or an empty string if not found.
}
\description{
Find an internal capsule directory
}
