% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_registry.R
\name{serious_internal_capsule_metadata}
\alias{serious_internal_capsule_metadata}
\title{Read metadata for an internal cell-based capsule}
\usage{
serious_internal_capsule_metadata(id, fallback)
}
\arguments{
\item{id}{Internal capsule id.}

\item{fallback}{Fallback metadata list.}
}
\value{
A metadata list.
}
\description{
Read metadata for an internal cell-based capsule
}
