% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_age_groups}
\alias{transform_age_groups}
\title{Transform age groups}
\usage{
transform_age_groups(
  data_tb,
  age_bands_lup,
  index_1L_chr,
  key_vars_chr = character(0),
  names_from_1L_chr,
  age_group_var_1L_chr = "AgeGroup",
  do_not_group_chr = character(0),
  rename_to_1L_chr = character(0),
  strict_1L_lgl = TRUE,
  values_to_1L_chr = "Value"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{age_bands_lup}{Age bands (a lookup table)}

\item{index_1L_chr}{Index (a character vector of length one)}

\item{key_vars_chr}{Key variables (a character vector), Default: character(0)}

\item{names_from_1L_chr}{Names from (a character vector of length one)}

\item{age_group_var_1L_chr}{Age group variable (a character vector of length one), Default: 'AgeGroup'}

\item{do_not_group_chr}{Do not group (a character vector), Default: character(0)}

\item{rename_to_1L_chr}{Rename to (a character vector of length one), Default: character(0)}

\item{strict_1L_lgl}{Strict (a logical vector of length one), Default: TRUE}

\item{values_to_1L_chr}{Values to (a character vector of length one), Default: 'Value'}
}
\value{
Data (a tibble)
}
\description{
transform_age_groups() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform age groups. The function returns Data (a tibble).
}
\keyword{internal}
