% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_output}
\alias{transform_output}
\title{Transform output}
\usage{
transform_output(output_ls)
}
\arguments{
\item{output_ls}{Output (a list)}
}
\value{
Output (a list)
}
\description{
transform_output() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform output. The function returns Output (a list).
}
\keyword{internal}
