% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_to_mdl_input}
\alias{transform_to_mdl_input}
\title{Transform to model input}
\usage{
transform_to_mdl_input(data_xx, ts_models_ls = make_ts_models_ls())
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{ts_models_ls}{Time series models (a list), Default: make_ts_models_ls()}
}
\value{
Data (a tsibble)
}
\description{
transform_to_mdl_input() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform to model input. The function returns Data (a tsibble).
}
\keyword{internal}
