% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_to_shorthand}
\alias{transform_to_shorthand}
\title{Transform to shorthand}
\usage{
transform_to_shorthand(
  data_tb,
  key_1L_chr = character(0),
  min_1L_int = 3L,
  original_xx = character(0),
  x_ready4show_correspondences = ready4show::ready4show_correspondences()
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{key_1L_chr}{Key (a character vector of length one), Default: character(0)}

\item{min_1L_int}{Minimum (an integer vector of length one), Default: 3}

\item{original_xx}{Original (an output object of multiple potential types), Default: character(0)}

\item{x_ready4show_correspondences}{PARAM_DESCRIPTION, Default: ready4show::ready4show_correspondences()}
}
\value{
Data (a tibble)
}
\description{
transform_to_shorthand() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform to shorthand. The function returns Data (a tibble).
}
\keyword{internal}
