% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_to_temporal}
\alias{transform_to_temporal}
\title{Transform to temporal}
\usage{
transform_to_temporal(
  X_Ready4useDyad = ready4use::Ready4useDyad(),
  metrics_chr,
  arrange_by_1L_chr = c("category", "name"),
  dictionary_r3 = ready4use::ready4use_dictionary(),
  index_1L_chr = "Date",
  key_vars_chr = character(0),
  temporal_vars_chr = make_temporal_vars()
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{metrics_chr}{Metrics (a character vector)}

\item{arrange_by_1L_chr}{Arrange by (a character vector of length one), Default: c("category", "name")}

\item{dictionary_r3}{Dictionary (a ready4 submodule), Default: ready4use::ready4use_dictionary()}

\item{index_1L_chr}{Index (a character vector of length one), Default: 'Date'}

\item{key_vars_chr}{Key variables (a character vector), Default: character(0)}

\item{temporal_vars_chr}{Temporal variables (a character vector), Default: make_temporal_vars()}
}
\value{
No return value, called for side effects.
}
\description{
transform_to_temporal() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform to temporal. The function is called for its side effects and does not return a value.
}
\keyword{internal}
