% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_for_price_year}
\alias{update_for_price_year}
\title{Update for price year}
\usage{
update_for_price_year(
  data_tb,
  cost_current_1L_chr = "Cost",
  cost_constant_1L_chr = "Cost",
  price_indices_dbl = numeric(0),
  price_ref_1L_int = 1L,
  time_var_1L_chr = "FiscalYear"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{cost_current_1L_chr}{Cost current (a character vector of length one), Default: 'Cost'}

\item{cost_constant_1L_chr}{Cost constant (a character vector of length one), Default: 'Cost'}

\item{price_indices_dbl}{Price indices (a double vector), Default: numeric(0)}

\item{price_ref_1L_int}{Price reference (an integer vector of length one), Default: 1}

\item{time_var_1L_chr}{Time variable (a character vector of length one), Default: 'FiscalYear'}
}
\value{
Data (a tibble)
}
\description{
update_for_price_year() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update for price year. The function returns Data (a tibble).
}
\keyword{internal}
