% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_medicare_data}
\alias{update_medicare_data}
\title{Update medicare data}
\usage{
update_medicare_data(
  medicare_tb,
  measures_chr = character(0),
  years_chr = character(0)
)
}
\arguments{
\item{medicare_tb}{Medicare (a tibble)}

\item{measures_chr}{Measures (a character vector), Default: character(0)}

\item{years_chr}{Years (a character vector), Default: character(0)}
}
\value{
Medicare (a tibble)
}
\description{
update_medicare_data() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update medicare data. The function returns Medicare (a tibble).
}
\keyword{internal}
