% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_retainers_ds}
\alias{update_retainers_ds}
\title{Update retainers dataset}
\usage{
update_retainers_ds(
  retainers_tb,
  cost_var_1L_chr = "Retainer amount",
  date_var_1L_chr = "Retainer date",
  end_date_dtm = NULL
)
}
\arguments{
\item{retainers_tb}{Retainers (a tibble)}

\item{cost_var_1L_chr}{Cost variable (a character vector of length one), Default: 'Retainer amount'}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Retainer date'}

\item{end_date_dtm}{End date (a date vector), Default: NULL}
}
\value{
Retainers (a tibble)
}
\description{
update_retainers_ds() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update retainers dataset. The function returns Retainers (a tibble).
}
\keyword{internal}
