% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_scenarios_tb}
\alias{update_scenarios_tb}
\title{Update scenarios tibble}
\usage{
update_scenarios_tb(
  scenarios_tb,
  base_case_1L_chr = "Base case",
  base_case_tb = NULL,
  change_var_1L_chr = "Total Cost",
  missing_1L_xx = NULL,
  difference_1L_int = 1L,
  outcomes_chr = "Budget Impact",
  scenario_1L_chr = "Scenario",
  tfmn_fn = NULL
)
}
\arguments{
\item{scenarios_tb}{Scenarios (a tibble)}

\item{base_case_1L_chr}{Base case (a character vector of length one), Default: 'Base case'}

\item{base_case_tb}{Base case (a tibble), Default: NULL}

\item{change_var_1L_chr}{Change variable (a character vector of length one), Default: 'Total Cost'}

\item{missing_1L_xx}{Missing length one (an output object of multiple potential types), Default: NULL}

\item{difference_1L_int}{Difference (an integer vector of length one), Default: 1}

\item{outcomes_chr}{Outcomes (a character vector), Default: 'Budget Impact'}

\item{scenario_1L_chr}{Scenario (a character vector of length one), Default: 'Scenario'}

\item{tfmn_fn}{Transformation (a function), Default: NULL}
}
\value{
Scenarios (a tibble)
}
\description{
update_scenarios_tb() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update scenarios tibble. The function returns Scenarios (a tibble).
}
\keyword{internal}
