% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_start_end_date}
\alias{update_start_end_date}
\title{Update start end date}
\usage{
update_start_end_date(
  end_dtm,
  start_dtm,
  index_day_1L_chr = "01",
  index_month_1L_chr = "07"
)
}
\arguments{
\item{end_dtm}{End (a date vector)}

\item{start_dtm}{Start (a date vector)}

\item{index_day_1L_chr}{Index day (a character vector of length one), Default: '01'}

\item{index_month_1L_chr}{Index month (a character vector of length one), Default: '07'}
}
\value{
New dates (a list)
}
\description{
update_start_end_date() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update start end date. The function returns New dates (a list).
}
\keyword{internal}
