% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_temporal_vars}
\alias{update_temporal_vars}
\title{Update temporal variables}
\usage{
update_temporal_vars(
  frequency_1L_chr = c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal"),
  temporal_vars_chr = make_temporal_vars()
)
}
\arguments{
\item{frequency_1L_chr}{Frequency (a character vector of length one), Default: c("daily", "weekly", "monthly", "quarterly", "yearly", "fiscal")}

\item{temporal_vars_chr}{Temporal variables (a character vector), Default: make_temporal_vars()}
}
\value{
Temporal variables (a character vector)
}
\description{
update_temporal_vars() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update temporal variables. The function returns Temporal variables (a character vector).
}
\keyword{internal}
