% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_to_full_tenure}
\alias{update_to_full_tenure}
\title{Update to full tenure}
\usage{
update_to_full_tenure(
  data_tb,
  date_var_1L_chr = "Date",
  drop_derived_1L_lgl = TRUE,
  end_date_dtm = NULL,
  start_date_dtm = NULL,
  tenure_var_1L_chr = "Tenure",
  uid_var_1L_chr = "UID",
  unit_1L_chr = "year"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Date'}

\item{drop_derived_1L_lgl}{Drop derived (a logical vector of length one), Default: TRUE}

\item{end_date_dtm}{End date (a date vector), Default: NULL}

\item{start_date_dtm}{Start date (a date vector), Default: NULL}

\item{tenure_var_1L_chr}{Tenure variable (a character vector of length one), Default: 'Tenure'}

\item{uid_var_1L_chr}{Unique identifier variable (a character vector of length one), Default: 'UID'}

\item{unit_1L_chr}{Unit (a character vector of length one), Default: 'year'}
}
\value{
Data (a tibble)
}
\description{
update_to_full_tenure() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update to full tenure. The function returns Data (a tibble).
}
\keyword{internal}
