% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_with_imputed}
\alias{update_with_imputed}
\title{Update with imputed}
\usage{
update_with_imputed(
  data_xx,
  impute_ls,
  lup_ls,
  args_ls = (list(NULL)),
  fns_ls = list(sample)
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{impute_ls}{Impute (a list)}

\item{lup_ls}{Lookup table list (a list of lookup tables)}

\item{args_ls}{Arguments (a list), Default: (
list(NULL)}

\item{fns_ls}{Functions (a list), Default: list(sample)}
}
\value{
Data (an output object of multiple potential types)
}
\description{
update_with_imputed() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update with imputed. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
