% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_builders.R
\name{apply_edges_to_steps}
\alias{apply_edges_to_steps}
\title{Apply graph edges to a list of steps}
\usage{
apply_edges_to_steps(steps, edges)
}
\arguments{
\item{steps}{List of \code{"serious_step"} objects.}

\item{edges}{Data frame with columns \code{from} and \code{to}.}
}
\value{
A list of updated \code{"serious_step"} objects.
}
\description{
Apply graph edges to a list of steps
}
