% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_cells.R
\name{capsule_cells}
\alias{capsule_cells}
\title{List cells in a capsule folder}
\usage{
capsule_cells(path)
}
\arguments{
\item{path}{Path to a capsule folder.}
}
\value{
A data frame describing the cells.
}
\description{
List cells in a capsule folder
}
