% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_cells.R
\name{capsule_get_cell}
\alias{capsule_get_cell}
\title{Get a cell from a capsule folder}
\usage{
capsule_get_cell(path, id)
}
\arguments{
\item{path}{Path to a capsule folder.}

\item{id}{Cell id.}
}
\value{
A list of class \code{"serious_cell"}.
}
\description{
Get a cell from a capsule folder
}
\seealso{
Other cell-based capsule API: 
\code{\link{capsule_add_cell}()},
\code{\link{capsule_cells}()},
\code{\link{capsule_connect_cells}()},
\code{\link{capsule_disconnect_cells}()},
\code{\link{capsule_move_cell}()},
\code{\link{capsule_set_cell_unlock}()},
\code{\link{capsule_update_cell}()}
}
\concept{cell-based capsule API}
