% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_cells.R
\name{capsule_set_cell_unlock}
\alias{capsule_set_cell_unlock}
\title{Set the unlock question of a cell}
\usage{
capsule_set_cell_unlock(
  path,
  id,
  question,
  expected_answer,
  case_sensitive = FALSE,
  success = NULL,
  failure = NULL
)
}
\arguments{
\item{path}{Path to a capsule folder.}

\item{id}{Cell id.}

\item{question}{Unlock question.}

\item{expected_answer}{Expected answer.}

\item{case_sensitive}{Logical. Should answer matching be case-sensitive?}

\item{success}{Optional success message.}

\item{failure}{Optional failure message.}
}
\value{
Invisibly returns the updated cell.
}
\description{
Set the unlock question of a cell
}
\seealso{
Other cell-based capsule API: 
\code{\link{capsule_add_cell}()},
\code{\link{capsule_cells}()},
\code{\link{capsule_connect_cells}()},
\code{\link{capsule_disconnect_cells}()},
\code{\link{capsule_get_cell}()},
\code{\link{capsule_move_cell}()},
\code{\link{capsule_update_cell}()}
}
\concept{cell-based capsule API}
