% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_cells.R
\name{capsule_update_cell}
\alias{capsule_update_cell}
\title{Update a cell in a capsule folder}
\usage{
capsule_update_cell(path, id, ...)
}
\arguments{
\item{path}{Path to a capsule folder.}

\item{id}{Cell id.}

\item{...}{Cell fields to update.}
}
\value{
Invisibly returns the updated cell.
}
\description{
Update a cell in a capsule folder
}
\seealso{
Other cell-based capsule API: 
\code{\link{capsule_add_cell}()},
\code{\link{capsule_cells}()},
\code{\link{capsule_connect_cells}()},
\code{\link{capsule_disconnect_cells}()},
\code{\link{capsule_get_cell}()},
\code{\link{capsule_move_cell}()},
\code{\link{capsule_set_cell_unlock}()}
}
\concept{cell-based capsule API}
