% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_dir.R
\name{check_capsule_dir}
\alias{check_capsule_dir}
\title{Check a SeRiouS capsule folder}
\usage{
check_capsule_dir(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to a capsule folder.}

\item{verbose}{Logical. If \code{TRUE}, prints diagnostic messages.}
}
\value{
Invisibly returns \code{TRUE} if the capsule folder is valid.
}
\description{
Check a SeRiouS capsule folder
}
