% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_registry.R
\name{get_capsule}
\alias{get_capsule}
\title{Get a registered SeRiouS capsule}
\usage{
get_capsule(id)
}
\arguments{
\item{id}{Capsule id.}
}
\value{
A \code{"learning_capsule"} object.
}
\description{
Get a registered SeRiouS capsule
}
