% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_medicare_data}
\alias{get_medicare_data}
\title{Get medicare data}
\usage{
get_medicare_data(path_1L_chr, clean_1L_lgl = FALSE)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}

\item{clean_1L_lgl}{Clean (a logical vector of length one), Default: FALSE}
}
\value{
Medicare (a tibble)
}
\description{
get_medicare_data() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get medicare data. The function returns Medicare (a tibble).
}
\keyword{internal}
