% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_dir.R
\name{load_capsule_dir}
\alias{load_capsule_dir}
\title{Load a SeRiouS capsule from a folder}
\usage{
load_capsule_dir(path)
}
\arguments{
\item{path}{Path to a capsule folder.}
}
\value{
A \code{"learning_capsule"} object.
}
\description{
Load a SeRiouS capsule from a folder
}
\seealso{
Other main user functions: 
\code{\link{available_capsules}()},
\code{\link{check_capsule_dir}()},
\code{\link{create_capsule_skeleton}()},
\code{\link{get_capsule}()},
\code{\link{run_capsule}()}
}
\concept{main user functions}
