% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_composite_forecast}
\alias{make_composite_forecast}
\title{Make composite forecast}
\usage{
make_composite_forecast(forecasts_tb)
}
\arguments{
\item{forecasts_tb}{Forecasts (a tibble)}
}
\value{
Forecasts (a tibble)
}
\description{
make_composite_forecast() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make composite forecast. The function returns Forecasts (a tibble).
}
\keyword{internal}
