% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_builders.R
\name{make_edges}
\alias{make_edges}
\title{Create capsule edges}
\usage{
make_edges(from, to)
}
\arguments{
\item{from}{Character vector of source step ids.}

\item{to}{Character vector of target step ids.}
}
\value{
A data frame with columns \code{from} and \code{to}.
}
\description{
Create capsule edges
}
