% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_builders.R
\name{make_layout}
\alias{make_layout}
\title{Create a manual capsule layout}
\usage{
make_layout(id, x, y)
}
\arguments{
\item{id}{Character vector of step ids.}

\item{x}{Numeric vector of x positions.}

\item{y}{Numeric vector of y positions.}
}
\value{
A data frame with columns \code{id}, \code{x}, and \code{y}.
}
\description{
Create a manual capsule layout
}
\seealso{
Other advanced capsule builders: 
\code{\link{build_capsule}()},
\code{\link{build_linear_capsule}()},
\code{\link{make_edges}()},
\code{\link{make_sections}()},
\code{\link{make_step}()},
\code{\link{validate_capsule}()}
}
\concept{advanced capsule builders}
