% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule_builders.R
\name{make_sections}
\alias{make_sections}
\title{Define capsule sections}
\usage{
make_sections(id, label = id, color = "#EEEEEE", border = "#999999")
}
\arguments{
\item{id}{Character vector of section identifiers.}

\item{label}{Character vector of section labels.}

\item{color}{Character vector of background colors.}

\item{border}{Character vector of border colors.}
}
\value{
A data frame describing capsule sections.
}
\description{
Define capsule sections
}
\seealso{
Other advanced capsule builders: 
\code{\link{build_capsule}()},
\code{\link{build_linear_capsule}()},
\code{\link{make_edges}()},
\code{\link{make_layout}()},
\code{\link{make_step}()},
\code{\link{validate_capsule}()}
}
\concept{advanced capsule builders}
