% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plateau.R
\name{mod_plateau_server}
\alias{mod_plateau_server}
\title{Plateau module server}
\usage{
mod_plateau_server(id, capsule, state)
}
\arguments{
\item{id}{Module id.}

\item{capsule}{A \code{"learning_capsule"} object.}

\item{state}{Shared reactive state.}
}
\value{
A Shiny module server.
}
\description{
Plateau module server
}
\keyword{internal}
