% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot forecast}
\usage{
plot_forecast(
  data_xx,
  ts_models_ls,
  metrics_chr = make_metric_vars(),
  x_label_1L_chr = ""
)
}
\arguments{
\item{data_xx}{Data (an output object of multiple potential types)}

\item{ts_models_ls}{Time series models (a list)}

\item{metrics_chr}{Metrics (a character vector), Default: make_metric_vars()}

\item{x_label_1L_chr}{X label (a character vector of length one), Default: ''}
}
\value{
Plot (an output object of multiple potential types)
}
\description{
plot_forecast() is a Plot function that plots data. Specifically, this function implements an algorithm to plot forecast. The function returns Plot (an output object of multiple potential types).
}
\keyword{internal}
