% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_serious.R, R/serious-package.R
\docType{package}
\name{serious-package}
\alias{serious-package}
\title{serious: Develop, Test and Apply Time Series Models for Mental Health Economic Analyses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for undertaking time series analysis of mental health services datasets developed with the ready4 framework (https://ready4-dev.github.io/ready4/). This development version of the serious package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for undertaking time series analysis of mental health services datasets developed with the ready4 framework (https://ready4-dev.github.io/ready4/). This development version of the serious package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/serious/}
  \item \url{https://github.com/ready4-dev/serious}
  \item \url{https://ready4-dev.github.io/ready4/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/serious/}
  \item \url{https://github.com/ready4-dev/serious}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID}) [copyright holder]

}
\keyword{internal}
